refactor: improve error message for version
On a fresh install of dctlenv, even when you install a version, it is not used by default so you have to run dctlenv use first, then run driftctl. So when you forget about this step, the error message shows "Failed to get version from dctlenv-version-name" which can be confusing as the user has no clue about what is dctlenv-version-name.